hacks for qemu skiboot w/ tpm
Normally the TPM is initialised in hostboot and there's an
event log to keep track of. In qemu case there is no hostboot,
so we need a bunch of ugly hacks instead.

The idea is to:

 - not try loading the CVC code from reserved memory. This isn't the
   fault of hostboot, just that we don't have a model of this code in
   qemu.

 - detect if we don't have an event log, and if not, assume we're
   in qemu and create an empty one with the TSS.

 - run the tpm init command if we're in qemu mode.

It's all super hacky, but it's enough to get trusted boot PCR extensions to work.

To extend to secure boot, you'd need to copy some of the hacks over.

Signed-off-by: Daniel Axtens <dja@axtens.net>